requirements.txt
PyMuPDF
sentence-transformers
requests
pandas
python-dotenv
tiktoken
fastapi
uvicorn
streamlit
fastapi
openai
openpyxl
pymongo
celery
pandas
redis
python-multipart
websockets
PyJWT
python-jose

#dhdbe